Extract utils function
