+++ tools/auto-fuzz/manager.py
@@ -40,6 +40,7 @@ import oss_fuzz_manager
 import fuzz_driver_generation_python
 import fuzz_driver_generation_jvm
 import post_process
+import utils
 from classes import OSS_FUZZ_PROJECT
 
 from io import BytesIO
@@ -69,37 +70,6 @@ if not os.path.isdir(OSS_FUZZ_BASE):
     raise Exception("Could not find OSS-Fuzz directory")
 
 
-def get_next_project_folder(base_dir):
-    AUTOFUZZDIR = "autofuzz-"
-    max_idx = -1
-    for dirname in os.listdir(base_dir):
-        try:
-            idx = int(dirname.replace("autofuzz-", ""))
-            if idx > max_idx:
-                max_idx = idx
-        except:
-            pass
-    return os.path.join(base_dir, AUTOFUZZDIR + str(max_idx + 1))
-
-
-def run_cmd(cmd, timeout_sec):
-    #print("Running command %s" % (cmd))
-    proc = subprocess.Popen(shlex.split(cmd),
-                            stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE)
-    timer = threading.Timer(timeout_sec, proc.kill)
-    try:
-        timer.start()
-        stdout, stderr = proc.communicate()
-        #print(stdout)
-        #print("---------")
-        #print(stderr)
-    finally:
-        no_timeout = timer.is_alive()
-        timer.cancel()
-    return no_timeout
-
-
 def run_static_analysis_python(git_repo, oss_fuzz_base_project,
                                base_oss_fuzz_project_dir):
     basedir = oss_fuzz_base_project.project_folder
@@ -124,49 +94,13 @@ def run_static_analysis_python(git_repo, oss_fuzz_base_project,
         "python3", FUZZ_INTRO_MAIN["python"], "--fuzzer", "../fuzz_1.py",
         "--package=%s" % (os.getcwd())
     ]
-    ret = run_cmd(" ".join(cmd), 1800)
+    ret = utils.run_cmd(" ".join(cmd), 1800)
     if not os.path.isfile("fuzzerLogFile-fuzz_1.data.yaml"):
         ret = False
     ret = True
 
     os.chdir(curr_dir)
     return ret
-
-
-def find_dir_build_type(dir):
-    """Determine the java build project type of the directory"""
-
-    if os.path.exists(os.path.join(dir, "pom.xml")):
-        return "maven"
-    elif os.path.exists(os.path.join(dir, "build.gradle")) or os.path.exists(
-            os.path.join(dir, "build.gradle.kts")):
-        return "gradle"
-    elif os.path.exists(os.path.join(dir, "build.xml")):
-        return "ant"
-    else:
-        return None
-
-
-def find_project_build_type(dir, proj_name):
-    # Search for current directory first
-    project_build_type = find_dir_build_type(dir)
-    if project_build_type:
-        return project_build_type
-
-    # Search for sub directory with name same as project name
-    for subdir in os.listdir(dir):
-        if os.path.isdir(os.path.join(dir, subdir)) and subdir == proj_name:
-            project_build_type = find_dir_build_type(os.path.join(dir, subdir))
-            if project_build_type:
-                return project_build_type
-
-    # Recursively look for subdirectory that contains build property file
-    for root, _, files in os.walk(dir):
-        project_build_type = find_dir_build_type(root)
-        if project_build_type:
-            return project_build_type
-
-    return None, None
 
 
 def build_jvm_project(oss_fuzz_base_project, base_oss_fuzz_project_dir,
@@ -319,61 +253,12 @@ def run_static_analysis_jvm(git_repo, oss_fuzz_base_project,
     return ret, jdk_base
 
 
-def copy_core_oss_fuzz_project_files(src_oss_project, dst_oss_project):
-    shutil.copy(src_oss_project.build_script, dst_oss_project.build_script)
-    shutil.copy(src_oss_project.project_yaml, dst_oss_project.project_yaml)
-    shutil.copy(src_oss_project.dockerfile, dst_oss_project.dockerfile)
-
-
-def copy_oss_fuzz_project_source(src_oss_project, dst_oss_project):
-    shutil.copytree(
-        os.path.join(src_oss_project.project_folder,
-                     src_oss_project.project_name),
-        os.path.join(dst_oss_project.project_folder,
-                     dst_oss_project.project_name))
-
-
 def tick_tqdm_tracker():
     global tqdm_tracker
     try:
         tqdm_tracker.update(1)
     except:
         return
-
-
-def cleanup_build_cache():
-    """Cleans up Docker build cache. This is needed becaus auto-fuzz builds
-    up a large docker build cache, which can take up hundreds of GBs on a
-    small run.
-    """
-    subprocess.check_call('docker builder prune --force',
-                          shell=True,
-                          stdout=subprocess.DEVNULL,
-                          stderr=subprocess.DEVNULL)
-
-
-def cleanup_base_directory(base_dir, project_name):
-    """Cleans up the base directory, removing unnecessary files after
-    the fuzzers auto-generation and checking process.
-    """
-    file_to_clean = [
-        'Fuzz.jar', 'Fuzz.class', 'ant.zip', 'gradle.zip', 'maven.zip',
-        'protoc.zip', 'jdk15.tar.gz', 'jdk17.tar.gz', 'jdk11.tar.gz',
-        'jdk8.tar.gz'
-    ]
-    dir_to_clean = [
-        'apache-maven-3.6.3', 'apache-ant-1.10.13', 'gradle-7.4.2',
-        'jdk-15.0.2', 'jdk-17', 'jdk-11.0.0.1', 'java-se-8u43-ri', 'protoc',
-        project_name, 'work/jar', 'work/proj', 'build-jar'
-    ]
-
-    for file in file_to_clean:
-        if os.path.isfile(os.path.join(base_dir, file)):
-            os.remove(os.path.join(base_dir, file))
-
-    for dir in dir_to_clean:
-        if os.path.isdir(os.path.join(base_dir, dir)):
-            shutil.rmtree(os.path.join(base_dir, dir))
 
 
 def build_and_test_single_possible_target(idx_folder,
@@ -402,8 +287,8 @@ def build_and_test_single_possible_target(idx_folder,
                                             language, benchmark)
 
     # Copy files from base OSS-Fuzz project
-    copy_core_oss_fuzz_project_files(oss_fuzz_base_project,
-                                     dst_oss_fuzz_project)
+    utils.copy_core_oss_fuzz_project_files(oss_fuzz_base_project,
+                                           dst_oss_fuzz_project)
     if language == "jvm":
         ant_path = os.path.join(oss_fuzz_base_project.project_folder,
                                 "ant.zip")
@@ -430,7 +315,7 @@ def build_and_test_single_possible_target(idx_folder,
         shutil.copy(protoc_path, protoc_dst)
         shutil.copytree(build_jar_path, build_jar_dst)
 
-    copy_oss_fuzz_project_source(oss_fuzz_base_project, dst_oss_fuzz_project)
+    utils.copy_oss_fuzz_project_source(oss_fuzz_base_project, dst_oss_fuzz_project)
 
     # Log dir
     idx_logdir = os.path.join(auto_fuzz_proj_dir, "autofuzz-log")
@@ -574,57 +459,12 @@ def run_builder_pool(autofuzz_base_workdir,
         pool.join()
 
         # Cleanup docker
-        cleanup_build_cache()
+        utils.cleanup_build_cache()
 
     try:
         tqdm_tracker.close()
     except:
         pass
-
-
-def copy_benchmark_project(base_dir, benchmark, language, destination):
-    shutil.copytree(os.path.join(base_dir, "benchmark", language, benchmark),
-                    destination)
-    return True
-
-
-def git_clone_project(github_url, destination):
-    cmd = ["git clone", github_url, destination]
-    try:
-        subprocess.check_call(" ".join(cmd),
-                              shell=True,
-                              timeout=600,
-                              stdout=subprocess.DEVNULL,
-                              stderr=subprocess.DEVNULL)
-    except subprocess.TimeoutExpired:
-        return False
-    except subprocess.CalledProcessError:
-        return False
-    return True
-
-
-def extract_class_list(projectdir):
-    """Extract a list of path for all java files exist in the project directory"""
-    project_class_list = []
-
-    for root, _, files in os.walk(projectdir):
-        for file in [file for file in files if file.endswith(".java")]:
-            path = os.path.join(root, file)
-            path = path.replace("%s/" % projectdir, "")
-            path = path.replace(".java", "").replace("/", ".")
-
-            # Filter some unrelated class
-            if "module-info" in path or "package-info" in path:
-                continue
-            if "test" in path or "Test" in path:
-                continue
-            if path.endswith("Exception"):
-                continue
-
-            if path not in project_class_list:
-                project_class_list.append(path)
-
-    return project_class_list
 
 
 def autofuzz_project_from_github(github_url,
@@ -639,7 +479,7 @@ def autofuzz_project_from_github(github_url,
     """
     print("Running autofuzz on %s" % (github_url))
     base_dir = os.getcwd()
-    autofuzz_base_workdir = get_next_project_folder(base_dir=base_dir)
+    autofuzz_base_workdir = utils.get_next_project_folder(base_dir=base_dir)
     os.mkdir(autofuzz_base_workdir)
 
     #autofuzz_data_dir = os.path.join(autofuzz_base_workdir, "base-autofuzz")
@@ -659,13 +499,13 @@ def autofuzz_project_from_github(github_url,
     # If benchmark option is true, instead of cloning the project from github,
     # copy the local benchmark directory of the chosen language instead.
     if benchmark:
-        if not copy_benchmark_project(
+        if not utils.copy_benchmark_project(
                 base_dir, github_url, language,
                 os.path.join(oss_fuzz_base_project.project_folder,
                              oss_fuzz_base_project.project_name)):
             return False
     else:
-        if not git_clone_project(
+        if not utils.git_clone_project(
                 github_url,
                 os.path.join(oss_fuzz_base_project.project_folder,
                              oss_fuzz_base_project.project_name)):
@@ -722,10 +562,10 @@ def autofuzz_project_from_github(github_url,
                                   oss_fuzz_base_project.project_name)
 
         # Extract class list for the target project
-        java_class_list = extract_class_list(projectdir)
+        java_class_list = utils.extract_class_list(projectdir)
 
         # Find project type
-        project_build_type = find_project_build_type(
+        project_build_type = utils.find_project_build_type(
             projectdir, oss_fuzz_base_project.project_name)
 
         if project_build_type:
@@ -780,8 +620,8 @@ def autofuzz_project_from_github(github_url,
             #                             parallelise=False,
             #                             dump_files=False)
         else:
-            cleanup_base_directory(base_oss_fuzz_project_dir,
-                                   oss_fuzz_base_project.project_name)
+            utils.cleanup_base_directory(base_oss_fuzz_project_dir,
+                                         oss_fuzz_base_project.project_name)
             return False
 
     # Check basic fuzzer and clean it afterwards
@@ -790,8 +630,8 @@ def autofuzz_project_from_github(github_url,
         OSS_FUZZ_BASE,
         log_dir=base_oss_fuzz_project_dir)
     if not res:
-        cleanup_base_directory(base_oss_fuzz_project_dir,
-                               oss_fuzz_base_project.project_name)
+        utils.cleanup_base_directory(base_oss_fuzz_project_dir,
+                                     oss_fuzz_base_project.project_name)
         return False
 
     # Generate all possible targets
@@ -839,8 +679,8 @@ def autofuzz_project_from_github(github_url,
 
     # Clean base directory for the project
     print("Cleaning base directory for %s" % (github_url))
-    cleanup_base_directory(base_oss_fuzz_project_dir,
-                           oss_fuzz_base_project.project_name)
+    utils.cleanup_base_directory(base_oss_fuzz_project_dir,
+                                 oss_fuzz_base_project.project_name)
 
     return True
 
@@ -931,16 +771,6 @@ def get_cmdline_parser() -> argparse.ArgumentParser:
     return parser
 
 
-def get_target_repos(targets, language):
-    if targets == "constants":
-        return constants.git_repos[language]
-    github_projects = []
-    with open(args.targets, 'r') as f:
-        for line in f:
-            github_projects.append(line.replace("\n", "").strip())
-    return github_projects
-
-
 if __name__ == "__main__":
     parser = get_cmdline_parser()
     args = parser.parse_args()
@@ -949,14 +779,14 @@ if __name__ == "__main__":
         if (args.benchmark):
             print("Benchmarking for python is not supported yet")
         else:
-            github_projects = get_target_repos(args.targets, 'python')
+            github_projects = utils.get_target_repos(args.targets, 'python')
             run_on_projects("python", github_projects, args.merge)
     elif args.language == 'java':
         if (args.benchmark):
             benchmark_projects = constants.benchmark['jvm']
             benchmark_on_projects("jvm", benchmark_projects, args.merge)
         else:
-            github_projects = get_target_repos(args.targets, 'jvm')
+            github_projects = utils.get_target_repos(args.targets, 'jvm')
             run_on_projects("jvm", github_projects, args.merge,
                             args.param_combination)
     else:
